Allow casts that don't change a constant's value

+++ src/cc1/expr.h
@@ -23,7 +23,7 @@ typedef struct consty
 			union
 			{
 				char *lbl;
-				int   memaddr;
+				unsigned long memaddr;
 			} bits;
 		} addr;
 	} bits;
